feat(setup): support ubuntu 24.04

+++ setup/setup__kubernetes.bash
@@ -1,19 +1,46 @@
 #!/bin/bash
 
-if [[ ! -f "/etc/apt/sources.list.d/kubernetes.list" ]]; then
+UBUNTU_VERSION=$(lsb_release -r -s)
+
+# Ubuntu < 24.04
+if [[ ! -f "/etc/apt/sources.list.d/kubernetes.list" && "$UBUNTU_VERSION" != "24.04" ]]; then
     sudo apt install -y apt-transport-https virtualbox virtualbox-ext-pack
     curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key add -
     cat <<EOF | sudo tee /etc/apt/sources.list.d/kubernetes.list
     deb https://apt.kubernetes.io/ kubernetes-xenial main
 EOF
+
     sudo apt update
 
     sudo apt install -y docker.io kubectl docker-compose parallel
     sudo apt-mark hold docker.io kubectl docker-compose
 
-    sudo usermod -aG docker $USER
+    sudo usermod -aG docker $HOME_USER_NAME && newgrp docker
 
     echo "{
+    \"registry-mirrors\": [ \"$HOME_REGISTRY_MIRROR\" ],
+    \"insecure-registries\": [\"$HOME_REGISTRY\"]
+}" | sudo tee /etc/docker/daemon.json
+
+    sudo systemctl daemon-reload
+    sudo systemctl restart docker
+fi
+
+# Ubuntu 24.04
+if [[ ! -f "/etc/apt/sources.list.d/kubernetes.list" && "$UBUNTU_VERSION" == "24.04" ]]; then
+    sudo apt-get install -y apt-transport-https ca-certificates virtualbox virtualbox-ext-pack curl
+    curl -fsSL https://pkgs.k8s.io/core:/stable:/v1.30/deb/Release.key | sudo gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
+    echo 'deb [signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://pkgs.k8s.io/core:/stable:/v1.30/deb/ /' | sudo tee /etc/apt/sources.list.d/kubernetes.list
+
+    sudo apt update
+
+    sudo apt install -y docker.io kubectl docker-compose parallel
+    sudo apt-mark hold docker.io kubectl docker-compose
+
+    sudo usermod -aG docker $HOME_USER_NAME && newgrp docker
+
+    echo "{
+    \"registry-mirrors\": [ \"$HOME_REGISTRY_MIRROR\" ],
     \"insecure-registries\": [\"$HOME_REGISTRY\"]
 }" | sudo tee /etc/docker/daemon.json
